requirements.txt
aiohttp
aiofiles
aiopathlib
fastapi
uvicorn[standard]
gunicorn
pydantic-settings==2.3.1
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.42.4
tabulate
tqdm
timm>=0.9.16
tiktoken>=0.7.0
accelerate>=0.30.1
sentencepiece>=0.2.0
sentence_transformers>=2.7.0
chatglm-cpp>=0.4.1
opencc-python-reimplemented==0.1.7
redis
motor
markdown
pygments
paramiko
openpyxl
pymupdf>=1.24.5
python-docx>=1.1.2
python-pptx>=0.6.23
jieba
ollama>=0.4.2
